+++ gluecodium/src/test/resources/smoke/nesting/output/cpp/src/smoke/OuterClassWithInheritance.cpp
@@ -1,0 +1,21 @@
+// -------------------------------------------------------------------------------------------------
+//
+//
+// -------------------------------------------------------------------------------------------------
+#include "smoke/OuterClassWithInheritance.h"
+namespace smoke {
+OuterClassWithInheritance::OuterClassWithInheritance() {
+    ::gluecodium::get_type_repository().add_type(this, "smoke_OuterClassWithInheritance");
+}
+OuterClassWithInheritance::~OuterClassWithInheritance() {
+    ::gluecodium::get_type_repository().remove_type(this);
+}
+OuterClassWithInheritance::InnerClass::InnerClass() {
+}
+OuterClassWithInheritance::InnerClass::~InnerClass() {
+}
+OuterClassWithInheritance::InnerInterface::InnerInterface() {
+}
+OuterClassWithInheritance::InnerInterface::~InnerInterface() {
+}
+}